requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
discord==1.7.3
discord.py==1.7.3
gevent==21.12.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pushbullet.py==0.12.0
python-dotenv==0.20.0
python-magic==0.4.27
requests==2.28.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wakeonlan==2.1.0
websocket==0.2.1
websocket-client==1.3.2
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
